requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-generativeai==0.3.2
openai==1.6.1
requests==2.31.0
python-dotenv==1.0.0